requirements.txt
telethon
pybit
fastapi
uvicorn
jinja2
python-dotenv
aiohttp